Variable data type change

+++ 0x13-bit_manipulation/5-flip_bits.c
@@ -13,7 +13,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	int count = 0;
-	int number = (n ^ m);
+	unsigned long int number = (n ^ m);
 
 	while (number)
 	{
